fix user with role and continue fix hasAnyRole in SecurityConfig and JwtAuthFilter also not recommend hasAnyAuthority

+++ src/main/java/_bbu/lawfirmapi/services/role/implement/RoleServiceImpl.java
@@ -3,13 +3,17 @@ package _bbu.lawfirmapi.services.role.implement;
 import _bbu.lawfirmapi.exceptions.NotFoundException;
 import _bbu.lawfirmapi.models.DTO.role.request.RoleRequest;
 import _bbu.lawfirmapi.models.DTO.role.response.RoleResponse;
+import _bbu.lawfirmapi.models.Entity.AppUser;
 import _bbu.lawfirmapi.models.Entity.Role;
 import _bbu.lawfirmapi.repositories.RoleRepository;
 import _bbu.lawfirmapi.services.role.RoleService;
 import lombok.RequiredArgsConstructor;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.Optional;
 
 @Service
 @RequiredArgsConstructor
@@ -18,37 +22,45 @@ public class RoleServiceImpl implements RoleService {
     private final RoleRepository roleRepository;
 //    private final RoleWithMyBatis roleWithMyBatis;
     // get all role method
+
+    public AppUser getCurrentUser(){
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        if(authentication == null || authentication.isAuthenticated() || authentication.getPrincipal().equals("anonymousUser")){
+            return null;
+        }
+        return (AppUser) authentication.getPrincipal();
+    }
     @Override
     public List<Role> getAllRoles(){
-        return roleRepository.findAll();
+        System.out.println("My Current User : " + getCurrentUser());
+        System.out.println(roleRepository.findAll().isEmpty());
+        return Optional.of(roleRepository.findAll())
+                .filter(list -> !list.isEmpty())
+                .orElseThrow(() -> new NotFoundException("None any role founded"));
     }
     @Override
     public Role findRoleByRoleId(Integer roleId){
-       return roleRepository.findById(roleId).orElseThrow(() -> new RuntimeException(" Role id not found"));
+       return roleRepository.findById(roleId).orElseThrow(() -> new NotFoundException(" Role id not found"));
     }
     @Override
     public RoleResponse createNewRoleList(RoleRequest newRoleRequest){
-        Role role = new Role();
-        role.setRoleName(newRoleRequest.getRoleName());
-        Role newRole = roleRepository.save(role);
-        return RoleResponse.builder()
-                .roleId(newRole.getRoleId())
-                .roleName(newRole.getRoleName())
-                .build();
+        Role newRole = newRoleRequest.toEntity();
+        newRole.setRoleName(newRoleRequest.getRoleName());
+        return roleRepository.save(newRole).toResponse();
     }
     @Override
     public RoleResponse updateRoleById(Integer roleId  , RoleRequest updateRole){
         Role updateNewRole = roleRepository.findById(roleId).orElseThrow(() -> new NotFoundException("role Id "  + roleId + "not found"));
         updateNewRole.setRoleName(updateRole.getRoleName());
-        Role savedRole = roleRepository.save(updateNewRole);
-        return RoleResponse.builder()
-                .roleId(savedRole.getRoleId())
-                .roleName(savedRole.getRoleName())
-                .build();
+
+        return roleRepository.save(updateNewRole).toResponse();
     }
     @Override
-    public void deleteRoleById(Integer roleId){
-         roleRepository.deleteById(roleId);
+    public void removeRoleById(Integer roleId){
+        if(roleRepository.findById(roleId).isEmpty()){
+            System.out.println("role deleted" + roleRepository.findById(roleId));
+            roleRepository.deleteById(roleId);
+        }
     }
 
 }
